Fix "ghost points" bug.
This happens where a map query is made with 'forks' disabled, meaning
only the latest version of each OSM element is returned.

In some cases an OSM 'way' that is excluded will reference OSM nodes
that are not present in the way that is ultimately returned. Prior to
this commit, those ways' unreferenced nodes would be returned
regardless, resulting in the final map being peppered with unlabeled
"ghost points", with no way referencing them.

This fixes the issue by tracking which nodes are referenced to by ways
that are excluded, and filtering them out of the final result.